♻️ Refactor(#6): js -> ts 변환
