+++ src/main.tsx
@@ -1,8 +1,8 @@
 import ReactDOM from 'react-dom/client';
 import './index.css';
 import { RouterProvider } from 'react-router-dom';
-import router from './routes/router.jsx';
+import router from './routes/router.tsx';
 
-ReactDOM.createRoot(document.getElementById('root')).render(
+ReactDOM.createRoot(document.getElementById('root')!).render(
   <RouterProvider router={router} />
 );
